requirements.txt
requests
websocket-client
more-itertools
py-cui
attrs
python-benedict
pytz
